add import task

+++ backend/models/_Base.js
@@ -49,7 +49,7 @@ class Model {
     }
     static findOrCreate(name, key, args) {
         let query = {};
-        query[key] = args.key;
+        query[key] = args[key];
         let result = jsondb.get(name).find(query).value();
         if (!result) {
             args.id = shortid.generate();
@@ -67,29 +67,27 @@ class Model {
         opt[queryKey] = args[queryKey];
         args.updatedAt = new Date();
         let result = jsondb.get(name).find(opt);
-        console.log(args.file);
         if (args.file && args.file.path) {
             let file = result.value().file;
             if (file && file.path) {
-                console.log(33333)
                 let fPath = path.join(__dirname, "../../src", file.path);
                 if (fs.existsSync(fPath)) fs.unlinkSync(fPath);
             }
-            console.log(22222)
             args.file.path = args.file.path.replace(path.join(__dirname, "../../src"), "");
         } else {
-            console.log("1111111");
             args.file = result.value().file;
         }
         result.assign(args).write();
         return result;
     }
     static delete(name, args = {}) {
-        let result = jsondb.get(name).find(args),
-            file = result.value().file;
-        if (file && file.path) {
-            let fPath = path.join(__dirname, "../../src", file.path);
-            if (fs.existsSync(fPath)) fs.unlinkSync(fPath);
+        let result = jsondb.get(name).find(args);
+        if (result.value()) {
+            let file = result.value().file;
+            if (file && file.path) {
+                let fPath = path.join(__dirname, "../../src", file.path);
+                if (fs.existsSync(fPath)) fs.unlinkSync(fPath);
+            }
         }
         result = jsondb.get(name).remove(args).write();
         return result;
